UDPv4 is sufficient! we always have 100.x overlay. IPv6 is additive
ipAllowed[0] is the v4

+++ client/internal/peer/conn_pq_test.go
@@ -17,14 +17,14 @@ type fakePQ struct {
 	ok  bool
 }
 
-func (f fakePQ) OfferPayload(string) ([]byte, int)          { return nil, 0 }
-func (f fakePQ) ShouldSendBootstrapOffer(string) bool       { return false }
-func (f fakePQ) AnswerPayload(string, []byte) ([]byte, int) { return nil, 0 }
-func (f fakePQ) OnAnswer(string, []byte)                    {}
-func (f fakePQ) PSK(string) (wgtypes.Key, bool)             { return f.psk, f.ok }
-func (f fakePQ) SetRemoteAddr(string, netip.AddrPort)       {}
-func (f fakePQ) OnDataPathRekeyed(string, time.Duration)    {}
-func (f fakePQ) OnDataPathDown(string)                      {}
+func (f fakePQ) OfferPayload(string) ([]byte, uint16)          { return nil, 0 }
+func (f fakePQ) ShouldSendBootstrapOffer(string) bool          { return false }
+func (f fakePQ) AnswerPayload(string, []byte) ([]byte, uint16) { return nil, 0 }
+func (f fakePQ) OnAnswer(string, []byte)                       {}
+func (f fakePQ) PSK(string) (wgtypes.Key, bool)                { return f.psk, f.ok }
+func (f fakePQ) SetRemoteAddr(string, netip.AddrPort)          {}
+func (f fakePQ) OnDataPathRekeyed(string, time.Duration)       {}
+func (f fakePQ) OnDataPathDown(string)                         {}
 
 // TestConn_presharedKey_PQ covers the post-quantum branch of presharedKey across the
 // three states that matter: a derived PSK is programmed, and — before one exists —
